requirements.txt
Django>=4.2.0
Pillow>=10.0.0
python-decouple>=3.8
dj-database-url>=2.1.0
gunicorn>=21.2.0
